refactor: extract the telegram-bot-serivce from the telegram-api-bot-service
