+++ src/libs/telegram-bot-service/index.ts
@@ -1,0 +1,1 @@
+export * from "./mocked-telegram-bot-service";
